GetMovies endpoint added

+++ backend/src/main/java/com/example/backend/dto/MovieResponse.java
@@ -4,7 +4,7 @@ import com.example.backend.model.Movie;
 import lombok.Builder;
 
 @Builder
-public record MovieResponse(Long id, String name, boolean isWatched, double rating) {
+public record MovieResponse(Long id, String name, boolean isWatched, int rating) {
     public static MovieResponse from(Movie movie) {
         return MovieResponse.builder().id(movie.getId()).name(movie.getName()).isWatched(movie.isWatched()).rating((movie.getRating() != null)? movie.getRating() : 0).build();
     }
